Rename module from_dataframe to _from_dataframe
