+++ pandas/api/interchange/__init__.py
@@ -3,6 +3,6 @@ Public API for DataFrame interchange protocol.
 """
 
 from pandas.core.interchange.dataframe_protocol import DataFrame
-from pandas.core.interchange.from_dataframe import from_dataframe
+from pandas.core.interchange._from_dataframe import from_dataframe
 
 __all__ = ["from_dataframe", "DataFrame"]
